% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{inefficiencyTerm.SFA}
\alias{inefficiencyTerm.SFA}
\title{predict function for SFA models}
\usage{
\method{inefficiencyTerm}{SFA}(object, estimator)
}
\description{
predict function for SFA models
}
