% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_fit.R
\name{panel.sfa.fit}
\alias{panel.sfa.fit}
\title{fit SFA models on panel data}
\usage{
panel.sfa.fit(y, X, K = NULL, ineff = -1L, spec = c("css90"), deb = F,
  debll = F)
}
\arguments{
\item{y}{- dependent variable (typically log of total product or cost)}

\item{X}{- explanatory variables (typically log of inputs or translog form)}

\item{K}{- matrix of panel data indeces}

\item{ineff}{- -1L/1L for production/cost inefficiency}

\item{spec}{- panel data model specification (currently only CSS, 1990)}

\item{deb}{- debug option}

\item{debll}{- debug option for log-likelohood functions}
}
\description{
fit SFA models on panel data
}
