% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFA.R
\name{print.summary.SFA}
\alias{print.summary.SFA}
\title{Print summary for SFA objects.}
\usage{
\method{print}{summary.SFA}(object, rounded = TRUE, ...)
}
\arguments{
\item{object}{object of class SFA}
}
\description{
Print summary for SFA objects.
}
\details{
Prints table of summary statistics for SFA model fitted by \code{sfa.fit} funtion.
}
