% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{rtnorm}
\alias{rtnorm}
\title{draw random values from truncated normal distribution}
\usage{
rtnorm(n = 1, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\description{
draw random values from truncated normal distribution
}

