% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFA.R
\name{sfa.fit}
\alias{sfa.fit}
\title{sfa.fit}
\usage{
sfa.fit(y, X, CM = NULL, CV = NULL, K = NULL, ineff = -1,
  structure = "cs", dist = "tnorm", spec = NULL, intercept = TRUE,
  intercept_CM = TRUE, intercept_CV = TRUE, sv_f = NULL, sv_cm = NULL,
  sv_cv = NULL, ll = NULL, opt_method = "SANN", opt_control = NULL,
  deb = F, debll = F)
}
\arguments{
\item{y}{dependent (production/cost) variable.}

\item{X}{variables of the production/cost function.}

\item{CM}{data for exogneous determinants of the conditional mean inefficiency.}

\item{CV}{data for exogneous determinants of the conditional variance of inefficiency term.}

\item{K}{matrix of panel data indeces.}

\item{ineff}{-1 (or 1) for production (or cost) function, where inefficiency decreases (or increases) the total output (or costs).}

\item{structure}{"cs" for cross-section or "panel" for panel data model. (not functional yet)}

\item{dist}{distribution of inefficiency term ("hnorm", "exp", "tnorm").}

\item{spec}{specifies what model of endogeneous inefficiency term should be used (currently only bc95 for cross-section implemented).}

\item{intercept}{TRUE/FALSE if the intercept term should be added to the main formula.}

\item{intercept_CM}{TRUE/FALSE if the intercept should be added to the conditional mean inefficiency formula.}

\item{intercept_CV}{TRUE/FALSE if the intercept should be added to the conditional inefficiency variance formula. (not functional yet)}

\item{sv_f}{starting values for frontier model parameters.}

\item{sv_cm}{starting values for conditional mean model parameters. (not functional yet)}

\item{sv_cv}{starting values for conditional variance model parameters. (not functional yet)}

\item{ll}{allows custom log-likelihood function that will be MINIMIZED.}

\item{opt_method}{optimization method.}

\item{opt_control}{list of options for optimization routine.}

\item{deb}{debug mode (TRUE/FALSE).}
}
\value{
Returns object of the class SFA which is a list object consisting:
\itemize{
\item coeff -- coefficients for stochastic frontier model
\item cm_coeff -- coefficients for conditional mean inefficiency model
\item residuals -- total residuals (= both u + v terms)
\item parameters -- vector of all parameters returend from miximization of log-likelihood
\item N -- total number of observations
\item ineff -- -1 (1 resp.) for production (cost resp.) function
\item ineff_name -- either "production" or "cost" string
\item data -- list of all data used for estimation (including unit vectors as intercepts if appropriate)
\item call is list of \itemize{
   \item intercept
   \item intercept_CM
   \item dist
   \item spec
   \item structure
}
\item loglik -- log-likehood
\item hessian -- hessian matrix
\item lmfit -- fitted linear model
}
}
\description{
Fits stochastic frontier analysis (SFA) model
}
\details{
Notice that the choice of optimization method may have significant impact on the results. It is recommanded to experiment with different optimization algorithms. Recommended are:
\itemize{
\item SANN -- In general, this is the most robust method. It can be slow with larger datasets or more complex models but the results tend to be better if parameters \code{maxit, tmax, temp} are set correctly (maxit > 1e4+, tmax = 15, temp = 1).
\item L-BFGS-B -- Fastest, but can crashes on complex models (infinite log-likelihood values etc...). If starting values are set well, it leads to the same results as SANN but much faster.
\item BFGS -- Fast, but can crashes on complex models (infinite log-likelihood values etc...).
}
See help for \code{optim()} function.
}
\examples{
See vignettes.

}

