% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAt.R
\name{SFA}
\alias{SFA}
\title{stochastic frontier analysis}
\usage{
SFA(formula, data = NULL, intercept = TRUE, intercept_CM = TRUE,
  dist = "hnormal", sv_f = NULL, par_mu = NULL, ineff = -1,
  opt_method = "BFGS", ...)
}
\description{
stochastic frontier analysis
}

