% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAt.R
\docType{package}
\name{SFAt-package}
\alias{SFAt}
\alias{SFAt-package}
\title{Package for Stochastic frontier analysis (SFA) using traditional statistical methods on (potentially heterogenous) cross-section and panel data}
\description{
This package provides a collection of methods for doing SFA (Stochastic frontier analysis) based on traditional statistical techniques.
It was inspired by some existing software for SFA (R packages frontier, sfa and Benchmarking, STATA commands frontier, xtfrontier and sfcross/sfpanel) but it aims to be
more comprehansive, modular/extendable and open (open source).
}
