% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_data_panel}
\alias{sim_data_panel}
\title{generate simulated panel data for testing purposes}
\usage{
sim_data_panel(k = 20, t = 10, x_mean = 10, x_sd = 1, z_mean = 10,
  z_sd = 2, x_coeff = c(10, 6, 3), z_intercept = 5, z_coeff = c(1.9,
  -0.9), sigma_u = 2, sigma_v = 3, ineff = -1, aslist = F)
}
\arguments{
\item{k}{number of individuals - cross section}

\item{t}{number of observations per individual}

\item{x_coeff}{coefficients for explanatory variables}

\item{z_coeff}{coefficients for exogeneous varaibles}

\item{sigma_u}{variance of the inefficiency term}

\item{sigma_v}{variance of the random noise}

\item{ineff}{production (-1) or cost inefficiency form (1),}

\item{aslist}{TRUE/FALSE to return list or array}
}
\description{
generate simulated panel data for testing purposes
}

