% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_css90.R
\name{fit_css90_model}
\alias{fit_css90_model}
\title{Cornwell, Schmidt, and Sickles (1990) model}
\usage{
fit_css90_model(y, X, K, ineff, deb = T)
}
\description{
Cornwell, Schmidt, and Sickles (1990) model
}

