% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict.SFA}
\alias{predict.SFA}
\title{predict function for SFA models}
\usage{
\method{predict}{SFA}(object, newdata = NULL, type = c("frontier",
  "efficiency", "inefficiency"), estimator = c("JLMS", "ME", "BC"))
}
\description{
predict function for SFA models
}
