% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAt.R
\name{sfa.fit}
\alias{sfa.fit}
\title{sfa.fit}
\usage{
sfa.fit(y, X, CM = NULL, CV = NULL, ineff = -1, structure = "cs",
  dist = "tnorm", spec = NULL, intercept = TRUE, intercept_CM = TRUE,
  intercept_CV = TRUE, sv_f = NULL, sv_cm = NULL, sv_cv = NULL,
  opt_method = "BFGS", deb = F, control_opt = NULL)
}
\arguments{
\item{y}{dependent (production/cost) variable.}

\item{X}{variables of the production/cost function.}

\item{CM}{data for exogneous determinants of the conditional mean inefficiency.}

\item{CV}{data for exogneous determinants of the conditional variance of inefficiency term.}

\item{ineff}{-1 (or 1) for production (or cost) function.}

\item{structure}{"cs" for cross-section or "panel" for panel data model.}

\item{dist}{distribution of inefficiency term ("hnorm", "exp", "tnorm").}

\item{spec}{specifies what model of endogeneous inefficiency term should be used (currently only bc95 for cross-section implemented).}

\item{intercept}{TRUE/FALSE if the intercept term should be added to the main formula.}

\item{intercept_CM}{TRUE/FALSE if the intercept should be added to the conditional mean inefficiency formula.}

\item{intercept_CV}{TRUE/FALSE if the intercept should be added to the conditional inefficiency variance formula.}

\item{sv_f}{starting value for frontier model parameters.}

\item{sv_cm}{starting value for conditional mean model parameters.}

\item{sv_cv}{starting value for conditional variance model parameters.}

\item{opt_method}{optimization method.}

\item{deb}{debug mode (TRUE/FALSE).}

\item{control_opt}{list of options for optimization routine.}
}
\value{
list
}
\description{
Fits stochastic frontier analysis (SFA) model
}

