% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAplus.R
\name{sfa.fit}
\alias{sfa.fit}
\title{stochastic frontier analysis}
\usage{
sfa.fit(y, X, Z = NULL, intercept = TRUE, intercept_Z = TRUE,
  model = "panel", dist = "tnorm", start_val = NULL, par_mu = NULL,
  form = "cost", opt_method = "BFGS", deb = T, ...)
}
\description{
stochastic frontier analysis
}

