% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAplus.R
\name{sfa}
\alias{sfa}
\title{stochastic frontier analysis}
\usage{
sfa(formula, data = NULL, intercept = TRUE, intercept_Z = TRUE,
  dist = "hnormal", start_val = NULL, par_mu = NULL, ineff = -1,
  opt_method = "BFGS", ...)
}
\description{
stochastic frontier analysis
}

