% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAt.R
\docType{package}
\name{SFAt-package}
\alias{SFAt}
\alias{SFAt-package}
\title{Package for doing Stochastic frontier analysis (SFA) based on traditional statistical methods on (potentially heterogenous) cross-section and panel data}
\description{
This package is a follow-up/was inspired by some earlier R packages for SFA (frontier, sfa, Benchmarking) and Stata frontier, xtfrontier and sfcross/sfpanel routines.
}

